+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 
 - Improved compatibilty of built-in DERP server with clients connecting over WebSocket.
+- Allow nodes to use SSH agent forwarding [#2145](https://github.com/juanfont/headscale/pull/2145)
 
 ## 0.23.0 (2024-09-18)